Initial cut of primitives from piet-scene
Pulls in the primitives from the piet-scene crate, replacing the geometry and path types with those from kurbo. There are a few alterations to naming a docs and the addition of some convenience "builder" methods to making constructing various types easier.

There are more types to add to this crate-- specifically things related to fonts and text. I though I'd keep the initial PR smallish, covering just enough to serve current piet-scene needs.

I have a working branch with updates to the full piet-gpu repo to support these changes that is ready to go after this lands.