Add font and image types
This adds a new `Blob` type that is used as a general shared buffer with a generated unique identifier and then builds types for font and image resources on top of that.

It also adds an `Image` variant to the `Brush` type and reworks gradients for better code reuse, cleaning up the API a bit.