Add Draw/DrawRef types to represent "fill or stroke"
In the spirit of `Brush` and `BrushRef` these are a pair of owned and reference types to abstractly represent either a fill or a stroke.

The motivation is for vello#284 where trying to parameterize on draw style within a builder that is consumed by `fill` or `stroke` methods is ergonomically painful.

I could add these types to vello but it probably makes more sense for them to live here.