Changes to Blob for multithreading and interop
This adds `Send + Sync` bounds to the data contained in the `Blob` type to enable usage in multithreaded scenarios.

Also adds `from_raw_parts` and `into_raw_parts` methods to the same type. This is to support interoperability with other crates (like cosmic-text) that don't directly depend on our resource types.