Bump kurbo and remove Stroke
Now that kurbo 0.10 can represent stroke styles, remove it from this crate. Note that this also removes the `scale` parameter, which was used to model the non-scaling stroke vector effect in SVG.

The `Fill` and `Style` enums are kept, as fill rule represents a rendering choice and not path geometry, and is not (at least at the moment) represented in kurbo.