Add `with_alpha` for `Brush`, `Gradient`, `Image` (#67)
There is now a `with_alpha` to go along with `multiply_alpha` that sets
the alpha to the given value. This makes it work the same as for colors
with the same sort of API.

While doing this, some code for multiplying alpha on a gradient was
moved to `Gradient` to make the interfaces the same on all things that
support this.

Fixes #51.