Add x/y extend modes and quality hint to images (#77)
Changes `Image` to support separate extend modes for each dimension.

Also adds an `ImageQuality` enum to allow the user to specify a hint for
rendering. I choose low/medium/high variants rather than specific
filters so that we're not making a specific commitment to any set.

Finally, renames `Format` to `ImageFormat` for consistency.