Separate out `ImageBrush` and `ImageBrushRef` from `Image` (#117)
The current `Image` type is a mix of fields that are inherent parts of
the image (data, format, width, height) and fields which apply to a
specific instance rendering of that image (extend_x, extend_y, quality,
alpha). This PR separates the latter out into a separate `ImageBrush`
type (and also an `ImageBrushRef<'a>` type mirroring `Brush` and
`BrushRef<'a>`).

The intended use here is that clients (Masonry, Blitz, etc) would store
a persistent `Image` and create an `ImageBrush` (or `ImageBrushRef` as
appropriate) on demand during rendering.

This is also intended to enable `Image` to be moved in the
`raw_resource_handle` crate and become a simple and stable interop type
along with `Blob` and `Font`.

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>